refactor: group views
