+++ lib/views/activity/activity_view.dart
