+++ lib/views/shares/shares_view.dart
