+++ lib/widgets/feed/comment.dart
@@ -3,7 +3,7 @@ import 'package:aimimi/constants/styles.dart';
 import 'package:aimimi/models/user.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:aimimi/views/comment_view.dart';
+import 'package:aimimi/views/activity/comment_view.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import "package:timeago/timeago.dart" as timeago;
