+++ lib/widgets/feed/comment_title.dart
@@ -1,6 +1,6 @@
 import 'package:aimimi/constants/styles.dart';
 import 'package:flutter/material.dart';
-import 'package:aimimi/views/comment_view.dart';
+import 'package:aimimi/views/activity/comment_view.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import "package:timeago/timeago.dart" as timeago;
