+++ lib/widgets/goal/goal_shared.dart
@@ -3,7 +3,7 @@ import 'package:aimimi/models/goal.dart';
 import 'package:aimimi/models/user.dart';
 import 'package:aimimi/services/auth_service.dart';
 import 'package:aimimi/services/goal_service.dart';
-import 'package:aimimi/views/goal_shared_view.dart';
+import 'package:aimimi/views/shares/goal_shared_view.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:provider/provider.dart';
